Move various utility functions from databass.util to databass.db.models as class/static methods
